requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-cors>=4.0.0
garminconnect>=0.2.0
python-dotenv>=1.0.0
cryptography>=41.0.0
apscheduler>=3.10.0
gunicorn>=21.0.0
